Adds backend for email sending to user contacts
The current schema of the commpanel email sending system allows for
sending to an arbitrary list of users. However, it could not send to an
arbitrary list of emails, so it was impossible to construct a commpanel
filter such as "email all guardians of all students enrolled in at least
one class".

This commit migrates MessageRequest to have a sendto_fn CharField. This
holds the module path of a function which takes a user and outputs a
list of related addresses to send to. The default behavior is the old
behavior, to only send to the user itself. However, additional functions
can be used instead specify guardian emails, emergency contact emails,
and combinations of those. During the processing of a MessageRequest,
each user in the PersistentQueryFilter of recipients is passed to the
sendto_fn, and those addresses are sent mail.

TODO: Frontend commpanel settings to utilize this feature.